More Tests for SpeakerBufferTest
git-svn-id: https://sovereigncomputing.net/svn/sneer/trunk@3037 ee3b4d21-14f5-0310-a1fa-c5d8b2925ec4

+++ sneer/bricks/src/sneer/skin/sound/PcmSoundPacket.java
@@ -13,14 +13,14 @@ public class PcmSoundPacket extends Tuple {
 
 	public final ImmutableByteArray payload;
 	
-	public final int sequence;
+	public final short sequence;
 
 	private PcmSoundPacket(ImmutableByteArray payload_, short sequence_) {
 		payload = payload_;
 		sequence = sequence_;
 	}
 
-	public PcmSoundPacket(PublicKey publisher, long publicationTime, ImmutableByteArray payload_, int sequence_) {
+	public PcmSoundPacket(PublicKey publisher, long publicationTime, ImmutableByteArray payload_, short sequence_) {
 		super(publisher, publicationTime);
 		payload = payload_;
 		sequence = sequence_;
